+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- `bit add` - exclude a component when its main file is excluded 
 - `bit test` - paint a summary table when testing multiple components 
 - reimplement cat-object command